+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Master
 * Interface changes
     * Remove exponential incrementer
     * Rename data! to json!
-    * Rename increments to steps
+    * Rename increments to steps & change default to 10
     * Remove linear
     * Add min_size
     * Add step_size